+++ examples/bitcoin-to-evm-fungible-transfer/src/transfer.p2tr.ts
@@ -1,19 +1,20 @@
 import {
   createBitcoinFungibleTransfer,
   TypeOfAddress,
+  getPublicKey,
+  broadcastTransaction,
+  fetchUTXOS,
+  getFeeEstimates,
+  processUtxos,
 } from "@buildwithsygma/bitcoin";
 import type { BitcoinTransferParams, UTXOData } from "@buildwithsygma/bitcoin";
 import { BIP32Factory } from "bip32";
-import { mnemonicToSeed } from "bip39";
-import { crypto, initEccLib, networks } from "bitcoinjs-lib";
-import { toXOnly } from "bitcoinjs-lib/src/psbt/bip371";
+import { initEccLib, networks, Signer } from "bitcoinjs-lib";
 import dotenv from "dotenv";
 import * as tinysecp from "tiny-secp256k1";
-import { broadcastTransaction, fetchUTXOS, getFeeEstimates, processUtxos } from '@buildwithsygma/utils'
 
-import {
-  calculateSize,
-} from "./blockstreamApi.js";
+import { calculateSize } from "./blockstreamApi.js";
+import { Environment } from "@buildwithsygma/core";
 
 dotenv.config();
 
@@ -29,7 +30,6 @@ const {
   ADDRESS,
   AMOUNT,
 } = process.env;
-
 
 if (
   !SOURCE_CAIPID ||
@@ -51,21 +51,22 @@ async function btcToEvmTransfer(): Promise<void> {
   initEccLib(tinysecp);
   const bip32 = BIP32Factory(tinysecp);
   console.log("Transfer BTC to EVM");
-  const seed = await mnemonicToSeed(MNEMONIC);
-  const rootKey = bip32.fromSeed(seed, networks.testnet);
-  const derivedNode = rootKey.derivePath(DERIVATION_PATH);
 
-  const publicKeyDropedDERHeader = toXOnly(derivedNode.publicKey);
+  const { tweakedSigner, publicKeyDropedDERHeader } = (await getPublicKey({
+    bip32,
+    mnemonic: MNEMONIC as string,
+    derivationPath: DERIVATION_PATH as string,
+    network: networks.testnet,
+    typeOfAddress: TypeOfAddress.P2TR,
+  })) as { publicKeyDropedDERHeader: Buffer; tweakedSigner: Signer };
 
-  const tweakedSigner = derivedNode.tweak(
-    crypto.taggedHash("TapTweak", publicKeyDropedDERHeader),
+  const feeRate = await getFeeEstimates(process.env.SYGMA_ENV, "5");
+  const utxos = await fetchUTXOS(
+    process.env.SYGMA_ENV,
+    ADDRESS as unknown as string
   );
 
-  const feeRate = await getFeeEstimates('5');
-  const utxos = await fetchUTXOS(
-    ADDRESS as unknown as string);
-
-  const processedUtxos = processUtxos(utxos, AMOUNT);
+  const processedUtxos = processUtxos(utxos, Number(BigInt(AMOUNT!)));
 
   const mapedUtxos = processedUtxos.map((utxo) => ({
     utxoTxId: utxo.txid,
@@ -79,7 +80,7 @@ async function btcToEvmTransfer(): Promise<void> {
     publicKey: publicKeyDropedDERHeader,
     depositAddress: ADDRESS as unknown as string,
     domainId: DESTINATION_CHAIN_ID,
-    amount: AMOUNT,
+    amount: BigInt(AMOUNT!),
     feeValue: BigInt(0),
     changeAddress: ADDRESS as unknown as string,
     signer: tweakedSigner,
@@ -87,11 +88,11 @@ async function btcToEvmTransfer(): Promise<void> {
   });
 
   const transferParams: BitcoinTransferParams = {
-    source: SOURCE_CAIPID,
+    source: SOURCE_CAIPID!,
     destination: DESTINATION_CHAIN_ID,
-    destinationAddress: DESTINATION_ADDRESS,
-    amount: AMOUNT,
-    resource: RESOURCE_ID,
+    destinationAddress: DESTINATION_ADDRESS!,
+    amount: BigInt(AMOUNT!),
+    resource: RESOURCE_ID!,
     utxoData: mapedUtxos,
     feeRate: BigInt(Math.ceil(feeRate)),
     publicKey: publicKeyDropedDERHeader,
@@ -99,6 +100,8 @@ async function btcToEvmTransfer(): Promise<void> {
     network: networks.testnet,
     changeAddress: ADDRESS,
     size: BigInt(size),
+    environment: Environment.TESTNET,
+    sourceAddress: "",
   };
 
   const transfer = await createBitcoinFungibleTransfer(transferParams);
@@ -114,8 +117,8 @@ async function btcToEvmTransfer(): Promise<void> {
   const tx = psbt.extractTransaction(true);
   console.log("Transaction hex", tx.toHex());
 
-  const txId = await broadcastTransaction(tx.toHex());
+  const txId = await broadcastTransaction(process.env.SYGMA_ENV, tx.toHex());
   console.log("Transaction broadcasted", `${EXPLORER_URL}/tx/${txId}`);
 }
 
-btcToEvmTransfer().finally(() => { });
+btcToEvmTransfer().finally(() => {});