+++ examples/evm-to-evm-erc1155-transfer/src/transfer.ts
@@ -1,11 +1,7 @@
+import { Eip1193Provider, getSygmaScanLink } from "@buildwithsygma/core";
 import {
-  Eip1193Provider,
-  Environment,
-  getSygmaScanLink,
-} from "@buildwithsygma/core";
-import {
-  createNonFungibleAssetTransfer,
-  NonFungibleTransferParams,
+  createSemiFungibleAssetTransfer,
+  SemiFungibleTransferParams,
 } from "@buildwithsygma/evm";
 import dotenv from "dotenv";
 import { Wallet, providers } from "ethers";
@@ -13,7 +9,7 @@ import Web3HttpProvider from "web3-providers-http";
 
 dotenv.config();
 
-const privateKey = process.env.PRIVATE_KEY;
+const privateKey = process.env.PRIVATE_KEY ?? "";
 
 if (!privateKey) {
   throw new Error("Missing environment variable: PRIVATE_KEY");
@@ -22,9 +18,9 @@ if (!privateKey) {
 const SEPOLIA_CHAIN_ID = 11155111;
 const CRONOS_TESTNET_CHAIN_ID = 338;
 const RESOURCE_ID =
-  "0x0000000000000000000000000000000000000000000000000000000000000200";
+  "0x0000000000000000000000000000000000000000000000000000000000000300";
 const SEPOLIA_RPC_URL =
-  process.env.SEPOLIA_RPC_URL || "https://eth-sepolia-public.unifra.io";
+  process.env.SEPOLIA_RPC_URL || "https://eth-sepolia.public.blastapi.io";
 
 const explorerUrls: Record<number, string> = {
   [SEPOLIA_CHAIN_ID]: "https://sepolia.etherscan.io",
@@ -34,33 +30,36 @@ const getTxExplorerUrl = (params: {
   chainId: number;
 }): string => `${explorerUrls[params.chainId]}/tx/${params.txHash}`;
 
-export async function erc721Transfer(): Promise<void> {
+export async function erc1155Transfer(): Promise<void> {
   const web3Provider = new Web3HttpProvider(SEPOLIA_RPC_URL);
   const ethersWeb3Provider = new providers.Web3Provider(web3Provider);
-  const wallet = new Wallet(privateKey ?? "", ethersWeb3Provider);
+  const wallet = new Wallet(privateKey, ethersWeb3Provider);
   const sourceAddress = await wallet.getAddress();
-  const destinationAddress = await wallet.getAddress();
+  const destinationAddress = await wallet.getAddress(); // Replace with the actual recipient address
 
-  const params: NonFungibleTransferParams = {
+  const params: SemiFungibleTransferParams = {
     source: SEPOLIA_CHAIN_ID,
     destination: CRONOS_TESTNET_CHAIN_ID,
     sourceNetworkProvider: web3Provider as unknown as Eip1193Provider,
     resource: RESOURCE_ID,
-    tokenId: process.env.TOKEN_ID as string,
+    tokenIds: [process.env.TOKEN_ID], // Ensure TOKEN_ID of NFT is set in your .env file
+    amounts: [BigInt(process.env.AMOUNT)], // Amount is always 1 for non-fungible tokens
     recipientAddress: destinationAddress,
     sourceAddress,
-    environment: process.env.SYGMA_ENV,
   };
 
-  const transfer = await createNonFungibleAssetTransfer(params);
-
+  const transfer = await createSemiFungibleAssetTransfer(params);
   const approvals = await transfer.getApprovalTransactions();
+
   console.log(`Approving Tokens (${approvals.length})...`);
   for (const approval of approvals) {
     const response = await wallet.sendTransaction(approval);
     await response.wait();
     console.log(
-      `Approved, transaction: ${getTxExplorerUrl({ txHash: response.hash, chainId: SEPOLIA_CHAIN_ID })}`
+      `Approved, transaction: ${getTxExplorerUrl({
+        txHash: response.hash,
+        chainId: SEPOLIA_CHAIN_ID,
+      })}`
     );
   }
 
@@ -68,8 +67,13 @@ export async function erc721Transfer(): Promise<void> {
   const response = await wallet.sendTransaction(transferTx);
   await response.wait();
   console.log(
-    `Depositted, transaction:  ${getSygmaScanLink(response.hash, process.env.SYGMA_ENV)}`
+    `Deposited, transaction: ${getSygmaScanLink(
+      response.hash,
+      process.env.SYGMA_ENV
+    )}`
   );
 }
 
-erc721Transfer().finally(() => {});
+erc1155Transfer().catch((error) => {
+  console.error("Error during ERC1155 transfer:", error);
+});