+++ packages/evm/src/fungibleAssetTransfer.ts
@@ -1,5 +1,11 @@
 import type { EthereumConfig, EvmResource } from '@buildwithsygma/core';
-import { Config, FeeHandlerType, ResourceType, SecurityModel } from '@buildwithsygma/core';
+import {
+  Config,
+  Environment,
+  FeeHandlerType,
+  ResourceType,
+  SecurityModel,
+} from '@buildwithsygma/core';
 import {
   Bridge__factory,
   ERC20__factory,
@@ -9,6 +15,7 @@ import { Web3Provider } from '@ethersproject/providers';
 import { BigNumber, constants, utils } from 'ethers';
 import type { ethers, PopulatedTransaction } from 'ethers';
 
+import { UnregisteredResourceHandlerError, UnsupportedResourceTypeError } from './errors.js';
 import { AssetTransfer } from './evmAssetTransfer.js';
 import type {
   EvmFee,
@@ -131,7 +138,7 @@ class FungibleAssetTransfer extends AssetTransfer {
 
   public setResource(resource: EvmResource): void {
     if (resource.type !== ResourceType.FUNGIBLE) {
-      throw new Error('Unsupported Resource type.');
+      throw new UnsupportedResourceTypeError(ResourceType.FUNGIBLE, resource.type);
     }
     this.transferResource = resource;
   }
@@ -270,13 +277,12 @@ export async function createFungibleAssetTransfer(
   params: FungibleTransferParams,
 ): Promise<FungibleAssetTransfer> {
   const config = new Config();
-  await config.init(process.env.SYGMA_ENV);
+  await config.init(params.environment ?? Environment.MAINNET);
 
   const transfer = new FungibleAssetTransfer(params, config);
 
   const isValid = await transfer.isValidTransfer();
-  if (!isValid)
-    throw new Error('Handler not registered, please check if this is a valid bridge route.');
+  if (!isValid) throw new UnregisteredResourceHandlerError(transfer.resource.resourceId);
 
   await transfer.setTransferAmount(params.amount);
   return transfer;