+++ packages/evm/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './approveAndCheckFns.js';
 export * from './balances.js';
 export * from './depositFn.js';
 export * from './transaction.js';
+export * from './assetTransferHelpers.js';